requirements.txt
# KANIDA.AI Terminal — Quant Intelligence Engine
# Core dependencies

# ── Web Framework ──────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# ── AI / Claude ────────────────────────────────────────────
anthropic>=0.28.0

# ── Data Sources ───────────────────────────────────────────
kiteconnect>=5.0.1          # Zerodha Kite Connect (NSE + F&O)
polygon-api-client>=1.13.0  # Polygon.io (US equities + options)

# ── Data Processing ────────────────────────────────────────
pandas>=2.2.0
numpy>=1.26.0

# ── Config / Env ───────────────────────────────────────────
python-dotenv>=1.0.1

# ── Scheduling ─────────────────────────────────────────────
apscheduler>=3.10.4

# ── Dev / Testing ──────────────────────────────────────────
pytest>=8.2.0
httpx>=0.27.0               # for FastAPI test client
